Integration of updated SIDENAV Api changes on 2.0-rc5. (Final)
