+++ src/main/java/gwt/material/design/demo/client/application/ApplicationView.ui.xml
@@ -47,7 +47,19 @@
   <g:HTMLPanel>
 
     <g:HTMLPanel ui:field="menu" />
+
     <m:MaterialContainer>
+        <m:MaterialPanel ui:field="titlePanel" backgroundColor="BLUE" textColor="WHITE"  marginTop="56" padding="56" shadow="1">
+            <m:MaterialPanel ui:field="panel">
+                <m:MaterialLabel ui:field="title" text="GWT-Material" fontSize="2.3em"/>
+                <m:MaterialLabel ui:field="description" text="A Material Design look and feel for GWT Apps plus Phonegap." fontSize="1.1em"/>
+                <m:MaterialPanel marginTop="20">
+                    <m:MaterialChip ui:field="chipXml" letter="X" textColor="WHITE" marginRight="12" text="XML" />
+                    <m:MaterialChip ui:field="chipJava" letter="J" textColor="WHITE" text="JAVA" marginRight="12" />
+                    <m:MaterialChip ui:field="chipSpecification" letter="S" textColor="WHITE" text="SPECIFICATION" />
+                </m:MaterialPanel>
+            </m:MaterialPanel>
+        </m:MaterialPanel>
       <g:HTMLPanel ui:field="main" addStyleNames="{style.panel} panel" />
 
       <m:MaterialPanel>
