+++ src/main/java/gwt/material/design/demo/client/application/menu/MenuView.java
@@ -64,54 +64,18 @@ class MenuView extends ViewWithUiHandlers<MenuUiHandlers> implements MenuPresent
     interface Binder extends UiBinder<Widget, MenuView> {
     }
 
-    private String link;
-    private String specification;
+
     private List<SearchObject> listSearches = new ArrayList<>();
 
     @UiField MaterialHeader header;
     @UiField MaterialNavBar navBar, navBarSearch;
     @UiField MaterialSideNav sideNav;
-    @UiField MaterialPanel panel, titlePanel;
-    @UiField MaterialLabel title, description;
     @UiField MaterialSearch txtSearch;
     @UiField MaterialComboBox<ThemeLoader.ThemeBundle> comboThemes;
-    @UiField MaterialChip chipXml, chipJava, chipSpecification;
 
     @Inject
     MenuView(Binder uiBinder) {
         initWidget(uiBinder.createAndBindUi(this));
-
-        chipJava.getElement().getStyle().setCursor(Style.Cursor.POINTER);
-        chipJava.addClickHandler(clickEvent -> {
-            String java = "https://github.com/GwtMaterialDesign/gwt-material-demo/tree/master/src/main/java/gwt/material/design/demo/client/application/" + link + ".java";
-            Window.open(java, "_blank", "");
-        });
-
-        chipXml.getElement().getStyle().setCursor(Style.Cursor.POINTER);
-        chipXml.addClickHandler(clickEvent -> {
-            String xml = "https://github.com/GwtMaterialDesign/gwt-material-demo/tree/master/src/main/java/gwt/material/design/demo/client/application/" + link + ".ui.xml";
-            Window.open(xml, "_blank", "");
-        });
-        chipSpecification.getElement().getStyle().setCursor(Style.Cursor.POINTER);
-        chipSpecification.addClickHandler(clickEvent -> {
-            Window.open(specification, "_blank", "");
-        });
-        ThemeManager.register(chipXml, ThemeManager.DARKER_SHADE);
-        ThemeManager.register(chipXml.getLetterMixin().getSpan(), ThemeManager.LIGHTER_SHADE);
-        ThemeManager.register(chipJava, ThemeManager.DARKER_SHADE);
-        ThemeManager.register(chipJava.getLetterMixin().getSpan(), ThemeManager.LIGHTER_SHADE);
-        ThemeManager.register(chipSpecification, ThemeManager.DARKER_SHADE);
-        ThemeManager.register(chipSpecification.getLetterMixin().getSpan(), ThemeManager.LIGHTER_SHADE);
-
-        sideNav.addOpeningHandler(event -> {
-            $(navBar).css("width","calc(100% - " + sideNav.getWidth() + "px)");
-            $(navBarSearch).css("width","calc(100% - " + sideNav.getWidth() + "px)");
-        });
-
-        sideNav.addClosingHandler(event -> {
-            $(navBar).css("width", "100%");
-            $(navBarSearch).css("width", "100%");
-        });
 
         // search close event
         txtSearch.addCloseHandler(event -> {
@@ -124,6 +88,15 @@ class MenuView extends ViewWithUiHandlers<MenuUiHandlers> implements MenuPresent
             navBarSearch.setVisible(true);
             navBar.setVisible(false);
         });
+
+        sideNav.addOpeningHandler(event -> {
+            $(header).css("width", "calc(100% - " + sideNav.getWidth() + "px)");
+        });
+
+        sideNav.addClosingHandler(event -> {
+            $(header).css("width", "100%");
+        });
+
         initThemes();
         initSearches();
     }
@@ -221,7 +194,6 @@ class MenuView extends ViewWithUiHandlers<MenuUiHandlers> implements MenuPresent
     protected void initThemes() {
         ThemeManager.initialize();
         ThemeManager.register(navBar, ThemeManager.DARKER_SHADE);
-        ThemeManager.register(titlePanel);
         buildThemeList();
         comboThemes.setSingleValue(ThemeManager.getBundle());
         comboThemes.addValueChangeHandler(event -> {
@@ -240,31 +212,6 @@ class MenuView extends ViewWithUiHandlers<MenuUiHandlers> implements MenuPresent
         comboThemes.addItem("Purple", ThemePurple.INSTANCE);
         comboThemes.addItem("Red", ThemeRed.INSTANCE);
         comboThemes.addItem("Yellow", ThemeYellow.INSTANCE);
-    }
-
-    @Override
-    public void setPageTitle(String title, String description, String link, String specification) {
-        this.title.setText(title);
-        this.description.setText(description);
-        this.link = link;
-        this.specification = specification;
-
-        if (link.isEmpty()) {
-            chipJava.setVisible(false);
-            chipXml.setVisible(false);
-        } else {
-            chipJava.setVisible(true);
-            chipXml.setVisible(true);
-        }
-
-        if (specification.isEmpty()) {
-            chipSpecification.setVisible(false);
-        } else {
-            chipSpecification.setVisible(true);
-        }
-
-        MaterialAnimator.animate(Transition.BOUNCEINLEFT, this.title, 1000);
-        MaterialAnimator.animate(Transition.BOUNCEINLEFT, this.description, 1000);
     }
 
     @UiHandler("btnSearch")
