--- src/main/java/gwt/material/design/demo/client/application/patterns/google/GoogleModule.java
@@ -1,31 +1,0 @@
-package gwt.material.design.demo.client.application.patterns.google;
-
-/*
- * #%L
- * GwtMaterial
- * %%
- * Copyright (C) 2015 - 2016 GwtMaterialDesign
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-
-
-import com.gwtplatform.mvp.client.gin.AbstractPresenterModule;
-
-public class GoogleModule extends AbstractPresenterModule {
-    @Override
-    protected void configure() {
-        bindPresenter(GooglePresenter.class, GooglePresenter.MyView.class, GoogleView.class, GooglePresenter.MyProxy.class);
-    }
-}
