--- src/main/java/gwt/material/design/demo/client/application/patterns/google/GoogleView.java
@@ -1,37 +1,0 @@
-package gwt.material.design.demo.client.application.patterns.google;
-
-/*
- * #%L
- * GwtMaterial
- * %%
- * Copyright (C) 2015 - 2016 GwtMaterialDesign
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-
-
-import com.google.gwt.uibinder.client.UiBinder;
-import com.google.gwt.user.client.ui.Widget;
-import com.google.inject.Inject;
-import com.gwtplatform.mvp.client.ViewImpl;
-
-public class GoogleView extends ViewImpl implements GooglePresenter.MyView {
-    public interface Binder extends UiBinder<Widget, GoogleView> {
-    }
-
-    @Inject
-    GoogleView(Binder uiBinder) {
-        initWidget(uiBinder.createAndBindUi(this));
-    }
-}
