--- src/main/java/gwt/material/design/demo/client/application/patterns/sidenav/SideNavView.ui.xml
@@ -1,28 +1,0 @@
-<!--
-  #%L
-  GwtMaterial
-  %%
-  Copyright (C) 2015 - 2016 GwtMaterialDesign
-  %%
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  
-       http://www.apache.org/licenses/LICENSE-2.0
-  
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-  #L%
-  -->
-<ui:UiBinder xmlns:ui='urn:ui:com.google.gwt.uibinder' xmlns:g='urn:import:com.google.gwt.user.client.ui'
-             xmlns:m="urn:import:gwt.material.design.client.ui">
-
-    <ui:with field='res' type='gwt.material.design.demo.client.resources.MaterialResources' />
-
-    <m:MaterialPanel ui:field="patternPanel">
-
-    </m:MaterialPanel>
-</ui:UiBinder>
