+++ src/main/java/gwt/material/design/demo/client/application/style/grid/GridView.ui.xml
@@ -200,7 +200,7 @@
     <m:MaterialRow addStyleNames="code">
       <m:MaterialBadge text="NEW TOOL" textColor="BLUE" shadow="1" layoutPosition="RELATIVE" float="RIGHT" backgroundColor="BLUE_LIGHTEN_5" />
       <m:MaterialTitle title="Resizer" description="An interactive viewer that helps designers test material design breakpoints across desktop, mobile, and tablet." />
-      <g:Frame url="https://material.io/resizer/#url=https%3A%2F%2Fgwtmaterialdesign.github.io%2Fgwt-material-patterns%2F" width="100%" height="700px" />
+      <g:Frame url="https://material.io/source/#url=https%3A%2F%2Fgwtmaterialdesign.github.io%2Fgwt-material-patterns%2F" width="100%" height="700px" />
     </m:MaterialRow>
   </g:HTMLPanel>
 </ui:UiBinder>
