+++ src/main/java/gwt/material/design/demo/client/application/templates/TemplatesView.ui.xml
@@ -1,4 +1,4 @@
-<!--
+    <!--
   #%L
   GwtMaterial
   %%
